+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+* Apply resources serially to avoid TLS timeout
+
 ## [1.8.6] - 2018-05-29
 ### Added
 * Key value pairs as yargs for apply function